report step2 complete

+++ app/src/main/java/jp/ac/uryukyu/ie/e235714/Hero.java
@@ -37,7 +37,7 @@ public class Hero extends LivingThing{
      */
 
      @Override
-    public void wounded(int damage){
+    public void wounded(double damage){
         double hitPoint = getHitPoint();
         setHitPoint(hitPoint-damage); 
         if( hitPoint < 0 ) {
